+++ R/PurpleAirQC_hourly_AB_00.R
@@ -7,16 +7,24 @@
 #' \code{pat_aggregate()}.
 #' @param min_count Aggregation bins with fewer than `min_count` measurements
 #' will be marked as `NA`.
+#' @param returnAllColumns Logical specifying whether to return all columns
+#' of statistical data generated for QC algorithm or just the final `pm25` 
+#' result.
 #'  
 #' @description Creates a \code{pm25} timeseries by averaging aggregated data
 #' from the A and B channels and applying the following QC logic:
 #' 
 #' \enumerate{
 #' \item{Create pm25 by averaging the A and B channel aggregation means}
-#' \item{Invalidate data where:  (min_count < 10)}
+#' \item{Invalidate data where:  (min_count < 20)}
 #' \item{No further QC}
 #' }
 #'
+#' @note Purple Air II sensors reporting after the June, 2019 firmware
+#' upgrade report data every 2 minutes or 30 measurements per hour. The default
+#' setting of \code{min_count = 20} is equivalent to a required data recovery
+#' rate of 67%.
+#' 
 #' @return Data frame with columns \code{datetime} and \code{pm25}.
 #' 
 #' @examples 
@@ -32,7 +40,8 @@
 
 PurpleAirQC_hourly_AB_00 <- function(
   aggregationStats,
-  min_count = 10
+  min_count = 20,
+  returnAllColumns = FALSE
 ) {
   
   # ----- hourly_AB_01 ---------------------------------------------------------
@@ -51,8 +60,13 @@ PurpleAirQC_hourly_AB_00 <- function(
       .data$pm25, 
       which(.data$min_count < min_count), 
       NA) 
-    ) %>%
-    dplyr::select(.data$datetime, .data$pm25)
+    )
+  
+  if ( !returnAllColumns ) {
+    hourlyData <- 
+      hourlyData %>%
+      dplyr::select(.data$datetime, .data$pm25)
+  }
   
   return(hourlyData)